+++ app/src/main/java/com/example/foodorderingapp/OrderAmount.java
@@ -1,15 +1,25 @@
 package com.example.foodorderingapp;
 
+import android.app.Notification;
+import android.app.NotificationChannel;
+import android.app.NotificationManager;
+import android.app.PendingIntent;
 import android.content.Intent;
+import android.os.Build;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.NotificationCompat;
 
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
+import com.google.firebase.database.ValueEventListener;
 import com.google.firebase.storage.FirebaseStorage;
 
 import java.text.DecimalFormat;
@@ -30,14 +40,16 @@ public class OrderAmount extends AppCompatActivity {
     Double totalCost=0.0;
     TextView itemSubtotal;
     TextView orderTotal, deliveryFee, tax;
-
+    private String loggedInUser;
 
     private void addToFirebase(){
         firebaseDbRef = FirebaseDatabase.getInstance().getReference().child("Order");
         Date currentTime = Calendar.getInstance().getTime();
         String id = firebaseDbRef.push().getKey();
-        order.setOrderedOn(currentTime.toString());
+        loggedInUser = ((MyApplication) this.getApplication()).getUserName();
+        createNotificationChannel();
 
+        order.setOrderedOn(currentTime.toString());
         order.setOrderedBy( ((MyApplication) this.getApplication()).getUserName());
         order.setHotelId(((MyApplication) this.getApplication()).getRestaurantName());
         order.setCompletionStatus("In Progress");
@@ -54,7 +66,15 @@ public class OrderAmount extends AppCompatActivity {
 
         orderedItems.deleteCharAt(orderedItems.length()-1);
         order.setOrderedItems(orderedItems.toString());
-        firebaseDbRef.child(id).setValue(order);
+        firebaseDbRef.child(id).setValue(order).addOnCompleteListener(new OnCompleteListener<Void>() {
+            @Override
+            public void onComplete(@NonNull Task task) {
+                if(task.isSuccessful()) {
+                    sendNotification();
+                }
+            }
+        });
+
         System.out.println("order NG"+order.getOrderedBy());
         System.out.println("order NG"+order.getOrderNo());
     }
@@ -88,4 +108,43 @@ public class OrderAmount extends AppCompatActivity {
         Intent intent = new Intent(this, MyOrdersListActivity.class);
         startActivity(intent);
     }
+
+    public void createNotificationChannel() {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            CharSequence name = getString(R.string.channel_name);
+            String description = getString(R.string.channel_description);
+            int importance = NotificationManager.IMPORTANCE_DEFAULT;
+            NotificationChannel channel = new NotificationChannel(getString(R.string.channel_id), name, importance);
+            channel.setDescription(description);
+            NotificationManager notificationManager = getSystemService(NotificationManager.class);
+            notificationManager.createNotificationChannel(channel);
+        }
+    }
+
+    public void sendNotification(){
+
+        Intent intent = new Intent(this, ReceiveNotificationActivity.class);
+        PendingIntent pIntent = PendingIntent.getActivity(this, (int) System.currentTimeMillis(), intent, PendingIntent.FLAG_IMMUTABLE);
+
+        Intent intentPresent = new Intent(this, MyOrdersListActivity.class);
+        PendingIntent pPresenetIntent = PendingIntent.getActivity(this, (int) System.currentTimeMillis(), intentPresent, PendingIntent.FLAG_IMMUTABLE);
+
+        String channelId = getString(R.string.channel_id);
+
+
+        Notification noti = new NotificationCompat.Builder(this,channelId)
+
+                .setContentTitle("Order Status")
+                .setContentText("Hey " +loggedInUser +" Your order is in progess")
+                .setSmallIcon(R.drawable.ic_launcher_foreground)
+                .setContentIntent(pIntent)
+                .addAction(R.drawable.gender_checked_background, "And more", pPresenetIntent).build();
+
+        NotificationManager notificationManager = (NotificationManager) getSystemService(NOTIFICATION_SERVICE);
+        noti.flags |= Notification.FLAG_AUTO_CANCEL ;
+
+        notificationManager.notify(0, noti);
+
+    }
+
 }
